+++ server/odc-service/src/main/java/com/oceanbase/odc/service/connection/logicaldatabase/LogicalTableUtils.java
@@ -17,10 +17,17 @@
 package com.oceanbase.odc.service.connection.logicaldatabase;
 
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeMap;
 import java.util.TreeSet;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -29,7 +36,13 @@ import java.util.stream.Collectors;
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
 
-import org.apache.hadoop.util.StringUtils;
+import org.apache.commons.collections4.CollectionUtils;
+
+import com.oceanbase.odc.common.lang.Pair;
+import com.oceanbase.odc.core.shared.Verify;
+import com.oceanbase.odc.core.shared.exception.UnexpectedException;
+import com.oceanbase.odc.service.connection.logicaldatabase.model.DataNode;
+import com.oceanbase.odc.service.connection.logicaldatabase.model.LogicalTable;
 
 /**
  * @Author: Lebie
@@ -37,83 +50,379 @@ import org.apache.hadoop.util.StringUtils;
  * @Description: []
  */
 public class LogicalTableUtils {
-    private static final String DELIMITER = ".";
+    private static final String PATTERN_PLACEHOLDER = "[#]";
+    private static final String PATTERN_PLACEHOLDER_REGEX = "\\[#\\]";
+
+    private static final String DOT_DELIMITER = ".";
+    private static final String COMMA_DELIMITER = ",";
     private static final Pattern DIGIT_PATTERN = Pattern.compile("\\d+");
 
+    public static List<LogicalTable> generatePatternExpressions(@NotEmpty List<DataNode> dataNodes,
+            @NotEmpty List<String> allDatabaseNames) {
+        // 找出所有逻辑表，并推导出 逻辑库表名的 databaseNamePattern 和 tableNamePattern
+        List<LogicalTable> logicalTables = identifyLogicalTables(dataNodes);
 
-
-    public static Map<String, List<String>> generatePatternExpressions(@NotEmpty List<String> fullTableNames, @NotEmpty List<String> allDatabaseNames) {
-        Map<String, List<String>> patternToPhysicalTablesMap = identifyBasicPatterns(fullTableNames);
-        Map<String, List<String>> patternToExpressionMap = new LinkedHashMap<>();
-        for (String pattern : patternToPhysicalTablesMap.keySet()) {
-            List<String> tableNames = patternToPhysicalTablesMap.get(pattern);
-            String expression = replacePlaceholdersWithRanges(pattern, tableNames);
-            patternToExpressionMap.put(expression, tableNames);
-        }
-        return patternToExpressionMap;
-    }
-
-
-    private static Map<String, List<String>> identifyBasicPatterns(@Valid @NotEmpty List<String> tableNames) {
-        // 先将表名分解为由数字序列和非数字序列组成的部分，构建基本模式
-        Map<String, List<String>> basePatternToTables = tableNames.stream().collect(
-            Collectors.groupingBy(tableName -> tableName.replaceAll(DIGIT_PATTERN.pattern(), "[#]"))
-        );
-        basePatternToTables.entrySet().removeIf(entry -> entry.getValue().size() == 1);
-
-        // 最终确定的模式到表的映射
-        Map<String, List<String>> finalPatterns = new LinkedHashMap<>();
-
-        basePatternToTables.forEach((basePattern, names) -> {
-            // 检查哪些位置的[#]应该被保留，哪些应该替换成实际的数字
-            String pattern = getConsistentNumberPattern(names);
-            finalPatterns.put(pattern, names);
-        });
-        return finalPatterns;
-    }
-
-    private static String replacePlaceholdersWithRanges(String pattern, List<String> tableNames) {
-        String placeholderRegex = "\\[#\\]";
-        String[] parts = pattern.split(placeholderRegex, -1);
-
-        List<Set<String>> numberSets = new ArrayList<>();
-        for (int i = 0; i < parts.length - 1; i++) {
-            numberSets.add(new TreeSet<>());
-        }
-
-        Pattern patternRegex = Pattern.compile(pattern.replaceAll(placeholderRegex, "(\\\\d+)"));
-        for (String tableName : tableNames) {
-            Matcher matcher = patternRegex.matcher(tableName);
-            if (matcher.matches()) {
-                for (int i = 0; i < parts.length - 1; i++) {
-                    // 直接保存数字的字符串形式
-                    numberSets.get(i).add(matcher.group(i + 1));
-                }
+        // 替换掉 pattern 中的 [#] 为具体的数字范围
+        logicalTables.stream().forEach(table -> {
+            String databaseNamePattern = table.getDatabaseNamePattern();
+            String tableNamePattern = table.getTableNamePattern();
+            // 既分库，又分表
+            if (databaseNamePattern.contains(PATTERN_PLACEHOLDER) && tableNamePattern.contains(PATTERN_PLACEHOLDER)) {
+                List<String> range = replacePlaceholdersWithRanges(table);
+                table.setFullNameExpression(String.join(COMMA_DELIMITER, range));
+            } else if (databaseNamePattern.contains(PATTERN_PLACEHOLDER)) {
+                // 只分库，不分表
+                List<String> range = replacePlaceholdersWithRanges(databaseNamePattern,
+                        table.getActualDataNodes().stream()
+                                .map(node -> node.getSchemaName())
+                                .collect(Collectors.toList()))
+                        .stream().map(r -> r + DOT_DELIMITER + tableNamePattern).collect(Collectors.toList());
+                table.setFullNameExpression(String.join(COMMA_DELIMITER, range));
+            } else if (tableNamePattern.contains(PATTERN_PLACEHOLDER)) {
+                // 只分表，不分库
+                List<String> range = replacePlaceholdersWithRanges(tableNamePattern,
+                        table.getActualDataNodes().stream()
+                                .map(node -> node.getTableName())
+                                .collect(Collectors.toList()))
+                        .stream().map(r -> databaseNamePattern + DOT_DELIMITER + r).collect(Collectors.toList());
+                table.setFullNameExpression(String.join(COMMA_DELIMITER, range));
             } else {
-                throw new IllegalArgumentException("表名 " + tableName + " 不符合模式 " + pattern);
+                throw new UnexpectedException(
+                        String.format("Unexpected pattern: %s.%s", databaseNamePattern, tableNamePattern));
+            }
+        });
+        return logicalTables;
+    }
+
+    private static List<String> replacePlaceholdersWithRanges(LogicalTable logicalTable) {
+        String databaseNamePattern = logicalTable.getDatabaseNamePattern();
+        String tableNamePattern = logicalTable.getTableNamePattern();
+        List<String> databaseNames = logicalTable.getActualDataNodes().stream()
+                .map(DataNode::getSchemaName)
+                .collect(Collectors.toList());
+        List<String> tableNames = logicalTable.getActualDataNodes().stream()
+                .map(DataNode::getTableName)
+                .collect(Collectors.toList());
+
+        // 如果每个分库的分表是完全一致的
+        if (hasSameTableNames(logicalTable)) {
+            List<String> tableNameExpression = replacePlaceholdersWithRanges(tableNamePattern, tableNames).stream()
+                    .map(r -> replaceRangeSign(r)).collect(Collectors.toList());
+            List<String> databaseNameExpression = replacePlaceholdersWithRanges(databaseNamePattern, databaseNames);
+            List<String> resultSegs = new ArrayList<>();
+            for (String databaseName : databaseNameExpression) {
+                for (String tableName : tableNameExpression) {
+                    resultSegs.add(databaseName + DOT_DELIMITER + tableName);
+                }
+            }
+            return resultSegs;
+        } else {
+            // 先按照每个分库分组，计算各自分组的表达式
+            Map<String, List<String>> databaseGroup = logicalTable.getActualDataNodes().stream()
+                    .collect(Collectors.groupingBy(DataNode::getSchemaName, () -> new TreeMap<>(createSchemaNameComparator(databaseNamePattern)),
+                            Collectors.mapping(DataNode::getTableName, Collectors.toList())));
+
+            if (canBeMerged(databaseGroup, databaseNamePattern, tableNamePattern)) {
+                List<String> databaseExpressions =
+                        replacePlaceholdersWithRanges(databaseNamePattern, new ArrayList<>(databaseGroup.keySet()));
+                Verify.singleton(databaseExpressions, "databaseExpressions.size");
+
+                List<String> tableExpressions = replacePlaceholdersWithRanges(tableNamePattern,
+                        databaseGroup.values().stream().flatMap(Collection::stream)
+                                .collect(Collectors.toList()));
+                Verify.singleton(tableExpressions, "tableExpressions.size");
+                return Arrays.asList(databaseExpressions.get(0) + DOT_DELIMITER + tableExpressions.get(0));
+            } else {
+                // 无法合并的情况，返回每个分组的表达式
+                Map<String, List<String>> databaseName2ExpressionSegs = databaseGroup.entrySet().stream()
+                        .collect(Collectors.toMap(Map.Entry::getKey,
+                                entry -> replacePlaceholdersWithRanges(tableNamePattern, entry.getValue()),
+                                (oldValue, newValue) -> oldValue, TreeMap::new));
+                return databaseName2ExpressionSegs.entrySet().stream()
+                        .map(entry -> entry.getValue().stream().map(r -> entry.getKey() + DOT_DELIMITER + r)
+                                .collect(Collectors.joining(COMMA_DELIMITER)))
+                        .collect(Collectors.toList());
             }
         }
+    }
 
-        StringBuilder result = new StringBuilder();
-        for (int i = 0; i < parts.length; i++) {
-            result.append(parts[i]);
-            if (i < numberSets.size()) {
-                result.append(generateRangeOrList(numberSets.get(i), pattern));
-            }
+    private static boolean canBeMerged(Map<String, List<String>> databaseGroup, String databaseNamePattern,
+            String tableNamePattern) {
+        // 可以合并的条件是，分库能被表示成 连续区间或步长的形式 且 各分库的表达式能按照分库的自然顺序组合成 连续区间或步长
+
+        // 每个分库下的表数量不完全一致，无法合并
+        if (!isSameValueSize(databaseGroup)) {
+            return false;
         }
 
-        // 如果发现不连续的数字序列，输出枚举的表名
-        if (result.toString().contains(",")) {
-            return String.join(",", tableNames);
+        List<String> databaseExpressions =
+                replacePlaceholdersWithRanges(databaseNamePattern, new ArrayList<>(databaseGroup.keySet()));
+        List<String> tableExpressions = replacePlaceholdersWithRanges(tableNamePattern,
+                databaseGroup.values().stream().flatMap(Collection::stream).collect(Collectors.toList()));
+
+        // 如果 分库表达式 或者 分表表达式 不能表示为一个表达式的形式，无法合并
+        if (databaseExpressions.size() != 1 || tableExpressions.size() != 1) {
+            return false;
         }
 
-        return result.toString();
+        // 如果每个序号都是按照自然序均分到每个库的，则可以合并
+        return isOrderByEachElement(extractNumbers(databaseGroup, tableNamePattern));
     }
 
 
-    private static String generateRangeOrList(Set<String> numberStrings, String patternPart) {
-        TreeSet<String> sortedNumberStrings = new TreeSet<>(numberStrings);
+    // 根据提供的正则表达式创建一个Comparator，实现类似 SQL order by 的效果
+    private static Comparator<String> createSchemaNameComparator(String pattern) {
+        String regex = pattern.replace(PATTERN_PLACEHOLDER_REGEX, "(\\\\d+)");
+        Pattern compiledPattern = Pattern.compile(regex);
 
+        return (schema1, schema2) -> {
+            Matcher matcher1 = compiledPattern.matcher(schema1);
+            Matcher matcher2 = compiledPattern.matcher(schema2);
+
+            if (!matcher1.matches() || !matcher2.matches()) {
+                return schema1.compareTo(schema2);
+            }
+
+            // 从左到右比较所有找到的数字
+            for (int i = 1; i <= Math.min(matcher1.groupCount(), matcher2.groupCount()); i++) {
+                int num1 = Integer.parseInt(matcher1.group(i));
+                int num2 = Integer.parseInt(matcher2.group(i));
+                if (num1 != num2) {
+                    return Integer.compare(num1, num2);
+                }
+            }
+
+            return schema1.compareTo(schema2);
+        };
+    }
+
+    private static List<List<String>> extractNumbers(Map<String, List<String>> databaseName2Tables,
+            String tablePattern) {
+        List<List<String>> allNumberLists = new ArrayList<>();
+        Pattern patternRegex = Pattern.compile(tablePattern.replaceAll(PATTERN_PLACEHOLDER_REGEX, "(\\\\d+)"));
+
+        for (Map.Entry<String, List<String>> entry : databaseName2Tables.entrySet()) {
+            List<String> tableNames = entry.getValue();
+            for (String tableName : tableNames) {
+                Matcher matcher = patternRegex.matcher(tableName);
+                List<String> numbers = new ArrayList<>();
+                if (matcher.matches()) {
+                    int groupCount = matcher.groupCount();
+                    for (int i = 1; i <= groupCount; i++) {
+                        numbers.add(matcher.group(i));
+                    }
+                }
+                allNumberLists.add(numbers);
+            }
+        }
+        return allNumberLists;
+    }
+
+    public static boolean isOrderByEachElement(List<List<String>> lists) {
+        if (lists == null || lists.size() < 2) {
+            // 如果列表为空或者只有一个子列表，我们认为它是有序的
+            return true;
+        }
+
+        for (int i = 1; i < lists.size(); i++) {
+            boolean isCurrentListBigger = false;
+            for (int j = 0; j < lists.get(i).size(); j++) {
+                int prev = Integer.parseInt(lists.get(i - 1).get(j));
+                int curr = Integer.parseInt(lists.get(i).get(j));
+                if (prev > curr) {
+                    return false; // 当前列的元素没有按照顺序排列
+                } else if (prev < curr) {
+                    isCurrentListBigger = true; // 这意味着不需要比较后续的列
+                    break;
+                }
+                // 如果 prev == curr, 我们需要继续比较下一列
+            }
+            // 如果在所有列中，所有元素都相等，我们将其视为按顺序排列
+            if (!isCurrentListBigger && i == lists.size() - 1) {
+                return true;
+            }
+        }
+        return true; // 所有的列表都正确排序
+    }
+
+    private static boolean isAllSorted(List<List<String>> numberStrings) {
+        for (List<String> strings : numberStrings) {
+            if (!isSorted(strings)) {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    private static boolean isSorted(List<String> numberStrings) {
+        if (numberStrings.size() < 2) {
+            return true; // 列表中有0个或1个元素，认为已排序
+        }
+        int previous = Integer.parseInt(numberStrings.get(0));
+        for (int i = 1; i < numberStrings.size(); i++) {
+            int current = Integer.parseInt(numberStrings.get(i));
+            if (previous > current) {
+                return false; // 如果前一个字符串大于当前字符串，则列表未排序
+            }
+            previous = current;
+        }
+        return true; // 所有相邻元素均已正确排序
+    }
+
+    private static boolean isSameValueSize(Map<String, List<String>> map) {
+        // 用于存储列表长度，初始化为-1表示尚未设置
+        int listLength = -1;
+
+        for (List<String> list : map.values()) {
+            if (listLength == -1) {
+                // 设置第一个列表的长度为参照标准
+                listLength = list.size();
+            } else if (list.size() != listLength) {
+                // 如果找到长度不一致的列表，返回false
+                return false;
+            }
+        }
+
+        // 所有列表长度一致
+        return true;
+    }
+
+    private static String replaceRangeSign(String expression) {
+        // 找到最后一个']'的位置
+        int lastCloseIndex = expression.lastIndexOf(']');
+        // 如果存在，则替换为']]
+        if (lastCloseIndex != -1) {
+            expression = expression.substring(0, lastCloseIndex) + "]]" + expression.substring(lastCloseIndex + 1);
+        }
+
+        // 找到最后一个'['的位置
+        int lastOpenIndex = expression.lastIndexOf('[');
+        // 如果存在，则替换为'[['
+        if (lastOpenIndex != -1) {
+            expression = expression.substring(0, lastOpenIndex) + "[[" + expression.substring(lastOpenIndex + 1);
+        }
+
+        return expression;
+    }
+
+    private static boolean hasSameTableNames(LogicalTable logicalTable) {
+        Map<String, List<String>> schemaName2TableNames = logicalTable.getActualDataNodes().stream()
+                .collect(Collectors.groupingBy(DataNode::getSchemaName, TreeMap::new,
+                        Collectors.mapping(DataNode::getTableName, Collectors.toList())));
+        // 使用SortedSet保证列表中的元素顺序
+        Set<SortedSet<String>> tableSets = new HashSet<>();
+        for (List<String> tables : schemaName2TableNames.values()) {
+            SortedSet<String> sortedTables = new TreeSet<>(tables);
+            tableSets.add(sortedTables);
+        }
+        return tableSets.size() == 1;
+    }
+
+    private static List<String> replacePlaceholdersWithRanges(String pattern, List<String> names) {
+        String[] parts = pattern.split(PATTERN_PLACEHOLDER_REGEX, -1);
+        int len = parts.length;
+        Verify.notLessThan(len, 2, "Pattern should contain at least one placeholder: " + pattern);
+        Verify.notGreaterThan(len, 3, "Pattern should contain at most two placeholders: " + pattern);
+
+        Pattern patternRegex = Pattern.compile(pattern.replaceAll(PATTERN_PLACEHOLDER_REGEX, "(\\\\d+)"));
+        // contains only one placeholder
+        if (len == 2) {
+            String range = generateRangeOrList(names.stream().map(name -> {
+                Matcher matcher = patternRegex.matcher(name);
+                if (matcher.matches()) {
+                    return matcher.group(1);
+                } else {
+                    throw new IllegalArgumentException("表名 " + name + " 不符合模式 " + pattern);
+                }
+            }).collect(Collectors.toSet()));
+            return Arrays.asList(parts[0] + range + parts[1]);
+        } else if (len == 3) {
+            // contains two placeholders
+            List<Pair<String, String>> numberPairs = names.stream().map(name -> {
+                Matcher matcher = patternRegex.matcher(name);
+                if (matcher.matches()) {
+                    return new Pair<>(matcher.group(1), matcher.group(2));
+                } else {
+                    throw new IllegalArgumentException("表名 " + name + " 不符合模式 " + pattern);
+                }
+            }).collect(Collectors.toList()).stream().sorted(Comparator.comparing(pair -> pair.left))
+                    .collect(Collectors.toList());
+
+            Pair<String, String> rangePair = generateRangeOrList(numberPairs);
+            // 笛卡尔积的情况
+            if (rangePair != null) {
+                return Arrays.asList(parts[0] + rangePair.left + parts[1] + rangePair.right + parts[2]);
+            } else {
+                List<String> result = new ArrayList<>();
+                // 不是笛卡尔积的情况，需要按照第一个 [#] 的数字来分组，在每个分组里，再对第二个 [#] 单独处理，然后将所有结果拼接起来
+                numberPairs.stream().collect(
+                        Collectors.groupingBy(pair -> pair.left, TreeMap::new, Collectors.toCollection(ArrayList::new)))
+                        .entrySet().forEach(entry -> {
+                            String rightRange = generateRangeOrList(
+                                    entry.getValue().stream().map(pair -> pair.right).collect(Collectors.toSet()));
+                            result.add(parts[0] + entry.getKey() + parts[1] + rightRange + parts[2]);
+                        });
+                return result;
+            }
+        } else {
+            throw new UnexpectedException("Unexpected pattern: " + pattern);
+        }
+    }
+
+
+    private static Pair<String, String> generateRangeOrList(List<Pair<String, String>> numberPairs) {
+        // 如果是笛卡尔积，则结果可以表示为 Pair<String, String> 的形式
+        // 并且 left 和 right 的表达式都遵循单个表达式的原则
+        if (isCartesianProduct(numberPairs)) {
+            Set<String> leftSet =
+                    numberPairs.stream().map(pair -> pair.left).collect(Collectors.toCollection(TreeSet::new));
+            Set<String> rightSet =
+                    numberPairs.stream().map(pair -> pair.right).collect(Collectors.toCollection(TreeSet::new));
+            return new Pair<>(generateRangeOrList(leftSet), generateRangeOrList(rightSet));
+        }
+        return null;
+    }
+
+    public static boolean isCartesianProduct(List<Pair<String, String>> pairs) {
+        Map<String, Set<String>> map = new HashMap<>();
+
+        // 将left值相同的right值收集到一个Set中
+        for (Pair<String, String> pair : pairs) {
+            String left = pair.left;
+            String right = pair.right;
+
+            // 使用HashSet来去重，并保存right值
+            if (!map.containsKey(left)) {
+                map.put(left, new HashSet<>());
+            }
+            map.get(left).add(right);
+        }
+
+        // 所有left键应该映射到相同尺寸的right值集合
+        Set<String> referenceSet = null; // 用第一个Set作为参照
+        for (Set<String> rightSet : map.values()) {
+            // 设置参照Set
+            if (referenceSet == null) {
+                referenceSet = rightSet;
+                continue;
+            }
+            // 如果有Set与参照不同，那么它不是笛卡尔积
+            if (!CollectionUtils.isEqualCollection(referenceSet, rightSet)) {
+                return false;
+            }
+        }
+        return true; // 所有Set都相同，确认是笛卡尔积
+    }
+
+    private static String generateRangeOrList(Set<String> numberStrings) {
+        if (numberStrings.size() == 1) {
+            return numberStrings.iterator().next();
+        }
+
+        // 这里这里，改排序方法！
+        // 创建一个TreeSet集合，使用自定义Comparator按数字大小对字符串进行排序
+        TreeSet<String> sortedNumberStrings = new TreeSet<>(Comparator.comparingInt(Integer::parseInt));
+        sortedNumberStrings.addAll(numberStrings);
         // 检查是否为连续序列
         boolean isConsecutive = true;
         String first = sortedNumberStrings.first();
@@ -122,9 +431,9 @@ public class LogicalTableUtils {
 
         for (String currentNumber : sortedNumberStrings.tailSet(first, false)) {
             int currentStep = Integer.parseInt(currentNumber) - Integer.parseInt(prevNumber);
-            if(step == Integer.MIN_VALUE) {
+            if (step == Integer.MIN_VALUE) {
                 step = currentStep; // 设置初始步长
-            } else if(currentStep != step) {
+            } else if (currentStep != step) {
                 isConsecutive = false; // 发现非连续的步长
                 break;
             }
@@ -144,11 +453,38 @@ public class LogicalTableUtils {
                 return String.format("[%s-%s:%d]", start, end, step);
             }
         } else {
-            // 不连续的数字，输出全部表名
-            return String.join(",", sortedNumberStrings);
+            // 不连续的数字，返回空字符串，这种情况，调用方需要自行处理，比如拼接全部表名
+            return String.format("[%s]", String.join(COMMA_DELIMITER, numberStrings));
         }
     }
 
+    private static List<LogicalTable> identifyLogicalTables(@Valid @NotEmpty List<DataNode> dataNodes) {
+        // 先将表名分解为由数字序列和非数字序列组成的部分，构建基本模式
+        Map<String, List<DataNode>> basePattern2Tables = dataNodes.stream().collect(
+                Collectors.groupingBy(dataNode -> dataNode.getTableName().replaceAll(DIGIT_PATTERN.pattern(), "[#]")));
+        basePattern2Tables.entrySet().removeIf(entry -> entry.getValue().size() == 1);
+
+        // 最终确定的模式到表的映射
+        Map<String, List<DataNode>> finalPatterns = new LinkedHashMap<>();
+
+        basePattern2Tables.forEach((basePattern, nodes) -> {
+            // 检查哪些位置的[#]应该被保留，哪些应该替换成实际的数字
+            String pattern =
+                    getConsistentNumberPattern(nodes.stream().map(DataNode::getTableName).collect(Collectors.toList()));
+            finalPatterns.put(pattern, nodes);
+        });
+
+        List<LogicalTable> logicalTables = finalPatterns.entrySet().stream().map(entry -> {
+            LogicalTable logicalTable = new LogicalTable();
+            logicalTable.setTableNamePattern(entry.getKey());
+            logicalTable.setActualDataNodes(entry.getValue());
+            logicalTable.setDatabaseNamePattern(getConsistentNumberPattern(logicalTable.getActualDataNodes().stream()
+                    .map(node -> node.getSchemaName())
+                    .collect(Collectors.toList())));
+            return logicalTable;
+        }).collect(Collectors.toList());
+        return logicalTables;
+    }
 
     private static String getConsistentNumberPattern(List<String> names) {
         // 转换为有数字组成的列表
@@ -178,11 +514,11 @@ public class LogicalTableUtils {
         }
 
         // 根据稳定的序列还原出最终的模式
-        StringBuilder patternBuilder = new StringBuilder(nameSections.get(0).get(0));  // 开始先添加第一部分的非数字序列
+        StringBuilder patternBuilder = new StringBuilder(nameSections.get(0).get(0)); // 开始先添加第一部分的非数字序列
 
         for (int idx = 1, varIdx = 0; idx < sectionCount; idx += 2, varIdx++) {
             if (isVariable.get(varIdx)) {
-                patternBuilder.append("[#]");
+                patternBuilder.append(PATTERN_PLACEHOLDER);
             } else {
                 patternBuilder.append(nameSections.get(0).get(idx));
             }
@@ -190,7 +526,6 @@ public class LogicalTableUtils {
                 patternBuilder.append(nameSections.get(0).get(idx + 1));
             }
         }
-
         return patternBuilder.toString();
     }
 }
