+++ server/odc-service/src/test/resources/logicaldatabase/logical_table_identification.yaml
@@ -496,3 +496,119 @@
           schema_name: "db7"
         - table_name: "tb_2024_02_t"
           schema_name: "db7"
+
+- data_nodes:
+    - table_name: "tb_0"
+      schema_name: "db0"
+    - table_name: "tb_1"
+      schema_name: "db0"
+    - table_name: "tb_2"
+      schema_name: "db1"
+    - table_name: "tb_3"
+      schema_name: "db1"
+  all_database_names:
+    - "db0"
+    - "db1"
+  logical_tables:
+    - table_name_pattern: "tb_[#]"
+      database_name_pattern: "db[#]"
+      full_name_expression: "db[0-1].tb_[0-3]"
+      actual_data_nodes:
+        - table_name: "tb_0"
+          schema_name: "db0"
+        - table_name: "tb_1"
+          schema_name: "db0"
+        - table_name: "tb_2"
+          schema_name: "db1"
+        - table_name: "tb_3"
+          schema_name: "db1"
+
+- data_nodes:
+    - table_name: "tb_0"
+      schema_name: "db_0_0"
+    - table_name: "tb_1"
+      schema_name: "db_0_1"
+    - table_name: "tb_2"
+      schema_name: "db_1_0"
+    - table_name: "tb_3"
+      schema_name: "db_1_1"
+  all_database_names:
+    - "db_0_0"
+    - "db_0_1"
+    - "db_1_0"
+    - "db_1_1"
+  logical_tables:
+    - table_name_pattern: "tb_[#]"
+      database_name_pattern: "db_[#]_[#]"
+      full_name_expression: "db_[0-1]_[0-1].tb_[0-3]"
+      actual_data_nodes:
+        - table_name: "tb_0"
+          schema_name: "db_0_0"
+        - table_name: "tb_1"
+          schema_name: "db_0_1"
+        - table_name: "tb_2"
+          schema_name: "db_1_0"
+        - table_name: "tb_3"
+          schema_name: "db_1_1"
+
+- data_nodes:
+    - table_name: "tb_0_0"
+      schema_name: "db_0"
+    - table_name: "tb_0_1"
+      schema_name: "db_1"
+    - table_name: "tb_1_0"
+      schema_name: "db_2"
+    - table_name: "tb_1_1"
+      schema_name: "db_3"
+  all_database_names:
+    - "db_0"
+    - "db_1"
+    - "db_2"
+    - "db_3"
+  logical_tables:
+    - table_name_pattern: "tb_[#]_[#]"
+      database_name_pattern: "db_[#]"
+      full_name_expression: "db_[0-3].tb_[0-1]_[0-1]"
+      actual_data_nodes:
+        - table_name: "tb_0_0"
+          schema_name: "db_0"
+        - table_name: "tb_0_1"
+          schema_name: "db_1"
+        - table_name: "tb_1_0"
+          schema_name: "db_2"
+        - table_name: "tb_1_1"
+          schema_name: "db_3"
+
+- data_nodes:
+    - table_name: "tb_0"
+      schema_name: "db_0"
+    - table_name: "tb_2"
+      schema_name: "db_0"
+    - table_name: "tb_4"
+      schema_name: "db_0"
+    - table_name: "tb_6"
+      schema_name: "db_1"
+    - table_name: "tb_8"
+      schema_name: "db_1"
+    - table_name: "tb_10"
+      schema_name: "db_1"
+  all_database_names:
+    - "db_0"
+    - "db_1"
+  logical_tables:
+    - table_name_pattern: "tb_[#]"
+      database_name_pattern: "db_[#]"
+      full_name_expression: "db_[0-1].tb_[0-10:2]"
+      actual_data_nodes:
+        - table_name: "tb_0"
+          schema_name: "db_0"
+        - table_name: "tb_2"
+          schema_name: "db_0"
+        - table_name: "tb_4"
+          schema_name: "db_0"
+        - table_name: "tb_6"
+          schema_name: "db_1"
+        - table_name: "tb_8"
+          schema_name: "db_1"
+        - table_name: "tb_10"
+          schema_name: "db_1"
